Move round function to package
